+++ src/full/Agda/Syntax/Internal/Generic.hs
@@ -157,7 +157,7 @@ instance TermLike EqualityView where
       <$> traverseTermM f t
     IdiomType t -> IdiomType
       <$> traverseTermM f t
-    EqualityType s eq l t a b -> EqualityType s eq
+    EqualityType r s eq l t a b -> EqualityType r s eq
       <$> traverse (traverseTermM f) l
       <*> traverseTermM f t
       <*> traverseTermM f a
@@ -166,7 +166,7 @@ instance TermLike EqualityView where
   foldTerm f = \case
     OtherType t -> foldTerm f t
     IdiomType t -> foldTerm f t
-    EqualityType s eq l t a b -> foldTerm f (l ++ [t, a, b])
+    EqualityType _r _s _eq l t a b -> foldTerm f (l ++ [t, a, b])
 
 -- | Put it in a monad to make it possible to do strictly.
 copyTerm :: (TermLike a, Monad m) => a -> m a